product_service/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
django-filter==23.3